remove the 8
