+++ c_models/makefiles/extra_neuron.mk
@@ -9,7 +9,7 @@ EXTRA_MODELS_DIR := $(abspath $(dir $(MAKEFILE_PATH))/../../)/
 BUILD_DIR := $(EXTRA_MODELS_DIR)c_models/build/$(APP)/
 
 # This is where the output .aplx files will go
-APP_OUTPUT_DIR := $(EXTRA_MODELS_DIR)python_models8/model_binaries/
+APP_OUTPUT_DIR := $(EXTRA_MODELS_DIR)python_models/model_binaries/
 # key for the database in this APP_OUTPUT_DIR
 # If you change APP_OUTPUT_DIR please use a lower case letter
 DATABASE_KEY = N
