+++ examples/my_example.py
@@ -4,23 +4,23 @@ from pyNN.utility.plotting import Figure, Panel
 import matplotlib.pyplot as plt
 
 # import models
-from python_models8.neuron.plasticity.stdp.timing_dependence\
+from python_models.neuron.plasticity.stdp.timing_dependence\
     .my_timing_dependence import (
         MyTimingDependence)
-from python_models8.neuron.plasticity.stdp.weight_dependence\
+from python_models.neuron.plasticity.stdp.weight_dependence\
     .my_weight_dependence import (
         MyWeightDependence)
-from python_models8.neuron.builds.my_model_curr_exp import MyModelCurrExp
-from python_models8.neuron.builds.my_full_neuron import MyFullNeuron
-from python_models8.neuron.builds.my_if_curr_exp_sEMD import MyIFCurrExpSEMD
-from python_models8.neuron.builds.my_model_curr_exp_my_input_type import (
+from python_models.neuron.builds.my_model_curr_exp import MyModelCurrExp
+from python_models.neuron.builds.my_full_neuron import MyFullNeuron
+from python_models.neuron.builds.my_if_curr_exp_sEMD import MyIFCurrExpSEMD
+from python_models.neuron.builds.my_model_curr_exp_my_input_type import (
     MyModelCurrExpMyInputType)
-from python_models8.neuron.builds.my_model_curr_my_synapse_type import (
+from python_models.neuron.builds.my_model_curr_my_synapse_type import (
     MyModelCurrMySynapseType)
-from python_models8.neuron.builds.my_model_curr_exp_my_additional_input \
+from python_models.neuron.builds.my_model_curr_exp_my_additional_input \
     import (
         MyModelCurrExpMyAdditionalInput)
-from python_models8.neuron.builds.my_model_curr_exp_my_threshold import (
+from python_models.neuron.builds.my_model_curr_exp_my_threshold import (
     MyModelCurrExpMyThreshold)
