+++ examples/my_sEMD_example.py
@@ -5,7 +5,7 @@ See https://www.cit-ec.de/en/nbs/spiking-insect-vision for more details
 
 # imports
 import pyNN.spiNNaker as p
-from python_models8.neuron.builds.my_if_curr_exp_sEMD import MyIFCurrExpSEMD
+from python_models.neuron.builds.my_if_curr_exp_sEMD import MyIFCurrExpSEMD
 
 import datetime
 from pyNN.utility.plotting import Figure, Panel
