+++ nmt_integration_tests/test_my_full_neuron.py
@@ -14,7 +14,7 @@
 
 import pyNN.spiNNaker as sim
 from .nwt_testbase import NwtTestBase
-from python_models8.neuron.builds.my_full_neuron import MyFullNeuron
+from python_models.neuron.builds.my_full_neuron import MyFullNeuron
 
 # Set the run time of the execution
 run_time = 1000
