+++ nmt_integration_tests/test_my_model_curr_exp.py
@@ -14,7 +14,7 @@
 
 import pyNN.spiNNaker as sim
 from .nwt_testbase import NwtTestBase
-from python_models8.neuron.builds.my_model_curr_exp import MyModelCurrExp
+from python_models.neuron.builds.my_model_curr_exp import MyModelCurrExp
 
 # Set the run time of the execution
 run_time = 1000
