+++ nmt_integration_tests/test_my_model_curr_my_synapse_type.py
@@ -14,7 +14,7 @@
 
 import pyNN.spiNNaker as sim
 from .nwt_testbase import NwtTestBase
-from python_models8.neuron.builds.my_model_curr_my_synapse_type import (
+from python_models.neuron.builds.my_model_curr_my_synapse_type import (
     MyModelCurrMySynapseType)
 
 # Set the run time of the execution
