+++ python_models/__init__.py
@@ -1,6 +1,6 @@
 import os
 from spynnaker.pyNN.data import SpynnakerDataView
-from python_models8 import model_binaries
+from python_models import model_binaries
 
 # This adds the model binaries path to the paths searched by sPyNNaker
 SpynnakerDataView.register_binary_search_path(
