+++ python_models/neuron/builds/my_if_curr_exp_sEMD.py
@@ -4,7 +4,7 @@ from spynnaker.pyNN.models.neuron.neuron_models import (
 from spynnaker.pyNN.models.neuron import AbstractPyNNNeuronModelStandard
 from spynnaker.pyNN.models.neuron.implementations import ModelParameter
 from spynnaker.pyNN.models.neuron.synapse_types import SynapseTypeExponential
-from python_models8.neuron.input_types.my_input_type_semd import (
+from python_models.neuron.input_types.my_input_type_semd import (
     MyInputTypeCurrentSEMD)
 from spynnaker.pyNN.models.neuron.threshold_types import ThresholdTypeStatic
