+++ python_models/neuron/builds/my_model_curr_exp_my_additional_input.py
@@ -2,22 +2,18 @@
 from spynnaker.pyNN.models.neuron import AbstractPyNNNeuronModelStandard
 
 # Components from main tools
-from spynnaker.pyNN.models.neuron.synapse_types.synapse_type_exponential\
-    import SynapseTypeExponential
-from spynnaker.pyNN.models.neuron.threshold_types.threshold_type_static\
-    import ThresholdTypeStatic
+from spynnaker.pyNN.models.neuron.input_types import InputTypeCurrent
+from spynnaker.pyNN.models.neuron.synapse_types import SynapseTypeExponential
+from spynnaker.pyNN.models.neuron.threshold_types import ThresholdTypeStatic
 from spynnaker.pyNN.models.defaults import default_initial_values
-
-# Additional components
-from python_models8.neuron.input_types.my_input_type \
-    import MyInputType
-from python_models8.neuron.neuron_models.my_neuron_model \
-    import MyNeuronModel
+from python_models.neuron.additional_inputs.my_additional_input import (
+    MyAdditionalInput)
+from python_models.neuron.neuron_models.my_neuron_model import MyNeuronModel
 
 
-class MyModelCurrExpMyInputType(AbstractPyNNNeuronModelStandard):
+class MyModelCurrExpMyAdditionalInput(AbstractPyNNNeuronModelStandard):
 
-    @default_initial_values({"v", "isyn_exc", "isyn_inh", "my_multiplicator"})
+    @default_initial_values({"v", "isyn_inh", "isyn_exc", "input_current"})
     def __init__(
             self,
 
@@ -35,9 +31,9 @@ class MyModelCurrExpMyInputType(AbstractPyNNNeuronModelStandard):
             isyn_exc=0.0,
             isyn_inh=0.0,
 
-            # input type parameters
-            my_multiplicator=0.0,
-            my_input_parameter=0.0):
+            # additional input parameters and state variables
+            my_additional_input_parameter=1.0,
+            input_current=0.0):
 
         # create neuron model class
         neuron_model = MyNeuronModel(i_offset, my_neuron_parameter, v)
@@ -47,20 +43,25 @@ class MyModelCurrExpMyInputType(AbstractPyNNNeuronModelStandard):
             tau_syn_E, tau_syn_I, isyn_exc, isyn_inh)
 
         # create input type model
-        input_type = MyInputType(my_multiplicator, my_input_parameter)
+        input_type = InputTypeCurrent()
 
         # create threshold type model
         threshold_type = ThresholdTypeStatic(v_thresh)
+
+        # create additional inputs
+        additional_input_type = MyAdditionalInput(
+            my_additional_input_parameter, input_current)
 
         # Create the model using the superclass
         super().__init__(
 
             # the model a name (shown in reports)
-            model_name="MyModelCurrExpMyInputType",
+            model_name="MyModelCurrExpMyAdditionalInput",
 
             # the matching binary name
-            binary="my_model_curr_exp_my_input_type.aplx",
+            binary="my_model_curr_exp_my_additional_input.aplx",
 
             # the various model types
             neuron_model=neuron_model, input_type=input_type,
-            synapse_type=synapse_type, threshold_type=threshold_type)
+            synapse_type=synapse_type, threshold_type=threshold_type,
+            additional_input_type=additional_input_type)
