+++ python_models/neuron/builds/my_model_curr_my_synapse_type.py
@@ -3,23 +3,22 @@ from spynnaker.pyNN.models.neuron import AbstractPyNNNeuronModelStandard
 
 # Components from main tools
 from spynnaker.pyNN.models.neuron.input_types import InputTypeCurrent
-from spynnaker.pyNN.models.neuron.synapse_types import SynapseTypeExponential
 from spynnaker.pyNN.models.neuron.threshold_types import ThresholdTypeStatic
-
-# Additional components
-from python_models8.neuron.neuron_models.my_neuron_model import MyNeuronModel
 from spynnaker.pyNN.models.defaults import default_initial_values
 
+# Additional components
+from python_models.neuron.neuron_models.my_neuron_model import MyNeuronModel
+from python_models.neuron.synapse_types.my_synapse_type import MySynapseType
 
-class MyModelCurrExp(AbstractPyNNNeuronModelStandard):
 
-    # Identify which of the values are state variables
-    @default_initial_values({"v", "isyn_exc", "isyn_inh"})
+class MyModelCurrMySynapseType(AbstractPyNNNeuronModelStandard):
+
+    @default_initial_values({"v", "my_exc_init", "my_inh_init"})
     def __init__(
             self,
 
             # neuron model parameters and state variables
-            my_neuron_parameter=0.0,
+            my_neuron_parameter=-70.0,
             i_offset=0.0,
             v=-70.0,
 
@@ -27,17 +26,18 @@ class MyModelCurrExp(AbstractPyNNNeuronModelStandard):
             v_thresh=-50.0,
 
             # synapse type parameters and state variables
-            tau_syn_E=5.0,
-            tau_syn_I=5.0,
-            isyn_exc=0.0,
-            isyn_inh=0.0):
+            my_ex_synapse_parameter=0.1,
+            my_in_synapse_parameter=0.1,
+            my_exc_init=0.0,
+            my_inh_init=0.0):
 
         # create neuron model class
         neuron_model = MyNeuronModel(i_offset, my_neuron_parameter, v)
 
         # create synapse type model
-        synapse_type = SynapseTypeExponential(
-            tau_syn_E, tau_syn_I, isyn_exc, isyn_inh)
+        synapse_type = MySynapseType(
+            my_ex_synapse_parameter, my_in_synapse_parameter,
+            my_exc_init, my_inh_init)
 
         # create input type model
         input_type = InputTypeCurrent()
@@ -45,14 +45,13 @@ class MyModelCurrExp(AbstractPyNNNeuronModelStandard):
         # create threshold type model
         threshold_type = ThresholdTypeStatic(v_thresh)
 
-        # Create the model using the superclass
         super().__init__(
 
             # the model a name (shown in reports)
-            model_name="MyModelCurrExp",
+            model_name="MyModelMySynapseType",
 
             # the matching binary name
-            binary="my_model_curr_exp.aplx",
+            binary="my_model_curr_my_synapse_type.aplx",
 
             # the various model types
             neuron_model=neuron_model, input_type=input_type,
