+++ python_models/neuron/input_types/my_input_type_semd.py
