+++ python_models/neuron/plasticity/stdp/timing_dependence/my_timing_dependence.py
