+++ python_models/neuron/plasticity/stdp/weight_dependence/my_weight_dependence.py
