+++ python_models/neuron/synapse_types/my_synapse_type.py
