+++ setup.cfg
@@ -14,7 +14,7 @@
 
 [metadata]
 name = sPyNNakerNewModelTemplate
-version = attr: python_models8._version.__version__
+version = attr: python_models._version.__version__
 description = SpiNNaker Template for New Models
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -48,8 +48,8 @@ install_requires =
 
 [options.packages.find]
 include =
-        python_models8
-        python_models8.*
+        python_models
+        python_models.*
 
 [options.package_data]
 * =
